configs_service/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
